refactor(core): scope exclude_row_ranges to pub(crate); extract shared indexed_row_ranges gap helper
Narrow exclude_row_ranges to pub(crate) (no external consumer). Extract the
index-manifest gap computation into global_index_build_common::indexed_row_ranges,
parameterized by index_type, so all global-index build builders share one copy
(mirrors Java's single indexedRowRanges). btree refactored to use it; behavior
unchanged.
